remove new private assembly references feature
This caused a wide array of mod errors and crashes due to common assemblies like 0Harmony.dll and Newtonsoft.Json.dll getting reloaded (even when not marked as a private assembly).

For repro steps and affected files, see this thread in #modded-tech-support on the Stardew Valley Discord:
> Content Patcher "Can't parse JSON file ... Unable to find a constructor" error after recent updates

We can revisit the feature in later versions.